Initial work, all tests passing (see caveat below)
I decided not to go the i18n route for now, since I wasn't happy with it
(and since it's not what DM users would expect). Instead, I just set
default_error_messages= to return the messages devise expects (with a
monkey-patch to ValidationErrors#add to allow it to accept a Symbol
argument).

There were four tests that weren't passing, but all for what I feel are
unimportant reasons, so I've overriden them in test/overrides_test.rb with
DM-specific expectations.

I'm not sure the method I've used for testing (as far as running devise's
testsuite) is ideal, but it works.