The Datamapper :required default can be changed, therefore schema.rb should explicty set it to false if not provided, since this is what devise expects.
DataMapper::Property.required(true)

I could not get the tests to run, nor find where this schema code is
tested. I also need this fix for my current app, which is why it is
applied to the v1.1.4 tag rather than master, since I do not want to
upgrade to devise 1.2.rc yet.