Add email_changed? method as expected by plataformatec/devise@69126a31
Related tests marked as pending because save! is apparently still running
validations. That may have something to do with https://github.com/datamapper/dm-validations/pull/13.
I think this only is an issue with invalid legacy data, so I haven't made
an effort to figure out exactly what the issue is.