Make configurable the overriding of ValidationErrors.default_error_messages
Most of the changes are only for the benefit of tests and this behavior
may not be desired, so allow for a different set of messages or disable
updating default_error_messages altogether.

For devise's benefit, the following messages should be added:
:not_found, :already_confirmed, :not_locked, and :expired
(Ideally, a merge would take care of just adding those that aren't defined
but dm-validations doesn't currently make default_error_messages readable,
so far as I can tell).